+++ src/Client/CosmosDbConfigurationException.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Client
+{
+    [FullyCovered]
+    public class CosmosDbConfigurationException : Exception
+    {
+        public CosmosDbConfigurationException(string message) : base(message)
+        {
+            
+        }
+    }
+}