+++ src/Client/EntityNotFoundException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Client
+{
+    [FullyCovered]
+    public class EntityNotFoundException : Exception
+    {
+        public EntityNotFoundException(string message) : base(message)
+        {
+        }
+    }
+}