+++ src/Client/FullyCoveredAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Client
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public sealed class FullyCoveredAttribute : Attribute
+    {
+        
+    }
+}