+++ src/Tests/TestObjectNoEntityType.cs
@@ -2,12 +2,11 @@
 
 namespace Tests
 {
-    internal class TestObject
+    internal class TestObjectNoEntityType
     {
         internal int Age { get; set; }
         internal List<string> Aliases { get; set; }
         internal string FirstName { get; set; }
         internal string Id { get; set; }
-        internal string EntityType { get; set; }
     }
 }